feat: rename dateExtent to timeExtent to match d3
Fixes #1111
