+++ packages/d3fc-extent/index.js
@@ -1,2 +1,3 @@
 export { default as extentLinear } from './src/linear';
+export { default as extentTime } from './src/time';
 export { default as extentDate } from './src/date';
